Speed up configuration loading
To speed up configuration loading somewhat, after reading and evaluating AUTODST.INI, the result is now saved in binary format in a new AUTODST.CFG file. This eliminates the need to completely parse the INI file the next time it is started; instead, the configuration is read directly from the small CFG file. It is only used when a change has been made to the INI file (i.e., the INI file is newer than the CFG file).